requirements.txt
pyramid
waitress
requests
python-dateutil